+++ src/DonorProfiles/resources/js/components/app/index.js
@@ -1,8 +1,0 @@
-const App = () => {
-	return (
-		<div>
-			<h1>Donor Profile App!</h1>
-		</div>
-	);
-};
-export default App;