+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 -   Donor Profile Donation History UI is now polished (#5566)
+-   Donor Profile edit UI now uses React Dropzone (#5563)
 
 ## 2.9.6 - 2021-01-13